Added Discord authentication and user upserting for new accounts

+++ src/routes/api/seed/+server.ts
@@ -1,8 +1,11 @@
-import { json } from "@sveltejs/kit";
+import { error, json } from "@sveltejs/kit";
 import { UserModel } from "$lib/models/User.model";
 import type { UserModelType } from "$lib/types/User.types";
+import { NODE_ENV } from "$env/static/private";
 
 export const GET = async() => {
+    if(NODE_ENV === 'production') throw error(403, 'Unable to seed database in production envrionment');
+
     await UserModel.deleteMany();
     await UserModel.insertMany([{
         discordId: '735281668566286417',
@@ -22,7 +25,9 @@ export const GET = async() => {
                 discordId: '1045326923775103027'
             }
         ],
-        following: []
+        following: [],
+
+        roles: []
     },
     {
 
@@ -43,7 +48,9 @@ export const GET = async() => {
                 avatarUrl: 'https://images-ext-1.discordapp.net/external/bippMZ8cXIbrZpIqxAUInNx9-CiuwYDC22qRfw__xWU/%3Fsize%3D512/https/cdn.discordapp.com/avatars/735281668566286417/a283626a6dd90e462b01d2f3c8680a11.png?format=webp&quality=lossless',
                 discordId: '735281668566286417'
             }
-        ]
+        ],
+
+        roles: []
     }] as Array<UserModelType>);
 
     return json({
